requirements.txt
mysqlclient
csv
sqlalchemy
pymysql
python-dotenv